requirements.txt
python-dotenv
pyttsx3
SpeechRecognition
streamlit
google-generativeai
pyaudio